+++ changelog/mediation/mobileads-unityads/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 4.12.1.0
+
+### Updated
+
+- Updated minimum supported `UnityAds` version to `4.12.1`
+- Updated minimum supported `YandexMobileAds` version to `7.2.0`
+
 ## Version 4.11.3.1
 
 ### Updated